refactor: split backend monoliths into focused packages
- auth.py (1,400 lines) → auth/ package (8 modules)
- background_tasks.py (803 lines) → tasks/ package (5 domain modules)
- 5 *_client.py files → clients/ package
- extract upload pipeline into 7 named phase functions
- extract shared tag operations to utilities/tags.py
- remove 2 loq.toml line-count exemptions

424 tests pass, lint clean. all public APIs preserved via __init__.py re-exports.
