+++ backend/src/backend/_internal/auth/__init__.py
@@ -1,0 +1,116 @@
+"""OAuth 2.1 authentication and session management."""
+
+from backend._internal.auth.account_groups import (
+    LinkedAccount,
+    PendingAddAccountData,
+    PendingScopeUpgradeData,
+    delete_pending_add_account,
+    delete_pending_scope_upgrade,
+    get_or_create_group_id,
+    get_pending_add_account,
+    get_pending_scope_upgrade,
+    get_session_group,
+    remove_account_from_group,
+    save_pending_add_account,
+    save_pending_scope_upgrade,
+    switch_active_account,
+)
+from backend._internal.auth.dependencies import (
+    get_optional_session,
+    require_artist_profile,
+    require_auth,
+)
+from backend._internal.auth.developer_tokens import (
+    DeveloperToken,
+    PendingDevTokenData,
+    delete_pending_dev_token,
+    get_pending_dev_token,
+    list_developer_tokens,
+    revoke_developer_token,
+    save_pending_dev_token,
+)
+from backend._internal.auth.encryption import _decrypt_data, _encrypt_data
+from backend._internal.auth.exchange import (
+    consume_exchange_token,
+    create_exchange_token,
+)
+from backend._internal.auth.oauth import (
+    check_artist_profile_exists,
+    ensure_artist_exists,
+    get_oauth_client,
+    get_oauth_client_for_scope,
+    get_public_jwks,
+    handle_oauth_callback,
+    start_oauth_flow,
+    start_oauth_flow_for_pds,
+    start_oauth_flow_with_scopes,
+)
+from backend._internal.auth.scopes import (
+    _check_scope_coverage,
+    _get_missing_scopes,
+    _parse_scopes,
+)
+from backend._internal.auth.session import (
+    CONFIDENTIAL_REFRESH_TOKEN_DAYS,
+    PUBLIC_REFRESH_TOKEN_DAYS,
+    Session,
+    create_session,
+    delete_session,
+    get_client_auth_method,
+    get_refresh_token_lifetime_days,
+    get_session,
+    is_confidential_client,
+    update_session_tokens,
+)
+
+__all__ = [
+    "CONFIDENTIAL_REFRESH_TOKEN_DAYS",
+    "PUBLIC_REFRESH_TOKEN_DAYS",
+    "DeveloperToken",
+    "LinkedAccount",
+    "PendingAddAccountData",
+    "PendingDevTokenData",
+    "PendingScopeUpgradeData",
+    "Session",
+    "_check_scope_coverage",
+    "_decrypt_data",
+    "_encrypt_data",
+    "_get_missing_scopes",
+    "_parse_scopes",
+    "check_artist_profile_exists",
+    "consume_exchange_token",
+    "create_exchange_token",
+    "create_session",
+    "delete_pending_add_account",
+    "delete_pending_dev_token",
+    "delete_pending_scope_upgrade",
+    "delete_session",
+    "ensure_artist_exists",
+    "get_client_auth_method",
+    "get_oauth_client",
+    "get_oauth_client_for_scope",
+    "get_optional_session",
+    "get_or_create_group_id",
+    "get_pending_add_account",
+    "get_pending_dev_token",
+    "get_pending_scope_upgrade",
+    "get_public_jwks",
+    "get_refresh_token_lifetime_days",
+    "get_session",
+    "get_session_group",
+    "handle_oauth_callback",
+    "is_confidential_client",
+    "list_developer_tokens",
+    "remove_account_from_group",
+    "require_artist_profile",
+    "require_auth",
+    "revoke_developer_token",
+    "save_pending_add_account",
+    "save_pending_dev_token",
+    "save_pending_scope_upgrade",
+    "start_oauth_flow",
+    "start_oauth_flow_for_pds",
+    "start_oauth_flow_with_scopes",
+    "switch_active_account",
+    "update_session_tokens",
+]
