+++ backend/src/backend/_internal/clients/__init__.py
@@ -1,0 +1,1 @@
+"""HTTP service clients."""
